feat: 增加独立FOV计算库；重构坐标类型。

+++ src/draw/draw_map.rs
@@ -9,8 +9,8 @@ use crate::tile_pos::TilePos;
 pub fn draw_map(ecs: &World, ctx: &mut GameContext) {
     let map = ecs.fetch::<Map>();
     for (map_idx, tile_type) in map.tiles.iter().enumerate() {
-        let x = (map_idx % TILE_WIDTH as usize) as u32;
-        let y = (map_idx / TILE_WIDTH as usize) as u32;
+        let x = map_idx as u32 % TILE_WIDTH;
+        let y = map_idx as u32 / TILE_WIDTH;
         let tile_pos = TilePos::new(x, y);
 
         let fg_color = if map.revealed_tiles[map_idx] {
